Add description to quiz sets; minor refactoring
