+++ assets/questions/en/wsbt_animation.json
@@ -1,5 +1,5 @@
 {
-  "name": "Animation 2",
+  "name": "Animation Intermediate",
   "description": "A set of advanced questions on animations",
   "author": "WSBT",
   "questions": [
