+++ assets/questions/zh/zh_zz.json
@@ -1,5 +1,5 @@
 {
-  "name": "动画2",
+  "name": "动画中等难度",
   "description": "从英文翻译而来的动画相关题库",
   "author": "zz",
   "questions": [
