--- lib/widget/multi_select_widget.dart
@@ -1,29 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../model/quiz_item.dart';
-import 'markdown.dart';
-
-class MultiSelectWidget extends StatelessWidget {
-  final List<Choice> items;
-  final void Function() onTap;
-
-  const MultiSelectWidget(
-      {super.key, required this.items, required this.onTap});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: items
-          .map((e) => CheckboxListTile(
-                controlAffinity: ListTileControlAffinity.leading,
-                value: e.selected,
-                onChanged: (checked) {
-                  onTap.call();
-                  e.selected = !e.selected;
-                },
-                title: Markdown(e.content),
-              ))
-          .toList(),
-    );
-  }
-}
